feat: remove the dependency of @monaco-editor/react
