+++ public/pages/Reindex/components/CreateIndexFlyout/CreateIndexFlyout.test.tsx
@@ -8,6 +8,7 @@ import React from "react";
 import CreateIndexFlyout from "./CreateIndexFlyout";
 import { coreServicesMock, browserServicesMock, apiCallerMock } from "../../../../../test/mocks";
 import { CoreServicesContext } from "../../../../components/core_services";
+import { ServicesContext } from "../../../../services";
 
 apiCallerMock(browserServicesMock);
 
@@ -15,7 +16,9 @@ describe("<CreateIndexFlyout /> spec", () => {
   it("renders the component", async () => {
     const component = render(
       <CoreServicesContext.Provider value={coreServicesMock}>
-        <CreateIndexFlyout sourceIndices={[]} commonService={browserServicesMock.commonService} onCloseFlyout={() => {}} />,
+        <ServicesContext.Provider value={browserServicesMock}>
+          <CreateIndexFlyout sourceIndices={[]} onCloseFlyout={() => {}} />,
+        </ServicesContext.Provider>
       </CoreServicesContext.Provider>
     );
     expect(component).toMatchSnapshot();
